ecommerce-backend/requirements.txt
sqlalchemy~=2.0.41
redis~=6.2.0
flask~=3.1.1
python-dotenv~=1.1.1
gunicorn~=23.0.0
mysql-connector-python~=9.3.0
faker~=37.4.0
